Introduce CallbackProxyObject
This class was inspired by ProxyTypes module created by Phillip J. Eby
https://pypi.python.org/pypi/ProxyTypes/

Sometimes, you may want a proxy’s subject to be determined dynamically
whenever the proxy is used. For this purpose, you can use the
CallbackProxyObject type, which accepts a callback function and creates
a proxy that will invoke the callback in order to get the target.